requirements.txt
Django==5.0.0
djangorestframework==3.14.0
pydicom==2.4.3
pynetdicom==2.0.2
Pillow>=10.2.0
numpy>=1.26.0
reportlab==4.0.7
python-dateutil==2.8.2
pytz==2023.3